+++ custom_components/rainpoint_local/const.py
@@ -15,6 +15,7 @@ PLATFORMS = [
     "binary_sensor",
     "button",
     "number",
+    "switch",
     "update",
     "valve",
 ]